Created a function named 'getComputerChoice'.
Created an array, 'const options', with the choices of "rock, paper,
and scissors".

Created a variable, 'const randomChoice', that will generate a random
number between 0 and 1, round it to the nearest integer, and assign it
to one of the options in the array.

Currently working on creating a function so the I can play against
the computer.