Player Choice & Computer Choice Function
Created a 'playRound' function with two parameters to
represent the player selection and the computer selection.

Made a bunch of 'if' statements for all the possible
match-ups.

Result prints properly in the console based on the
pre-determined player choice of "rock". Need to figure out
how to ask for a player input, next.