Creating Loop for Multiple Rounds
We need to create a loop that will play 5 Rounds
of rock, paper, scissors. Creating two variables, 'playerWins',
and 'computerWins', we set them to 0 for the start.

I'm understanding the concept of loops but still working
on full wrapping my mind around on how to create one
in different scenarios. Loops are such a cool but simple concept.

Next, working on the console will prompt the user
for their choice and how it will print to the console.