We have... ROCK, PAPER, SCISSORS!
I had to re-write many parts to my code. For example,
in the 'playRound' function I used 'const' for the 'playerSelection'
parameter. Const was not necessary as it is already
a function.

Originally I had 'compChoice' variable for 'computerSelection'
'compChoice' was set only once before the game started
so it would not change during the game. Instead I called
'getComputerChoice()' function inside the loop. This would
allow me to get a new computer choice after each round.

Will pause on this code until the upcoming lessons to build the GUI.